day2/module/requirements.txt
psycopg2==2.9.9
pandas==2.1.1
numpy==1.26.1
ipykernel==6.25.2
SQLAlchemy==2.0.22
python-dotenv==1.0.0
pymysql==1.1.0
pyarrow==14.0.0
click==8.1.7